[REFACTOR] View 파일 분리
